pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-logger"
version = "2.0.0"
description = "Log debugging tool for QA testers - analyze camera/hardware logs, detect issues, generate bug reports"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Alon Mozes"},
]

dependencies = [
    "Flask>=3.0.0",
    "Werkzeug>=3.0.1",
    "Flask-SQLAlchemy>=3.1.1",
    "SQLAlchemy>=2.0.23",
    "WTForms>=3.1.1",
    "Flask-WTF>=1.2.1",
    "python-dateutil>=2.8.2",
    "pandas>=2.1.4",
    "chardet>=5.2.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.0.7",
    "orjson>=3.9.10",
    "paramiko>=3.4.0",
    "boto3>=1.28.0",
    "anthropic>=0.75.0",
    "openai>=1.58.1",
    "requests>=2.31.0",
    "psutil>=5.9.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
server = ["gunicorn>=21.2.0"]

[project.scripts]
sentinel-logger = "main:main"

[tool.setuptools]
py-modules = ["main", "config"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.setuptools.package-data]
app = [
    "templates/**/*",
    "templates/*",
    "static/**/*",
    "static/*",
]
